chore(store, clients): Update player position data type to f32
Update player position data type to f32 for smoother movement and more accurate positioning.

This change updates the `Position` struct in the `store` crate to use f32 for x, y, and z coordinates. It also makes corresponding adjustments in the `clients` crate to handle the change in the data type.

This change allows for smoother movement and more accurate positioning of players in the game.

+++ store/src/player.rs
@@ -2,9 +2,9 @@ use serde::{ Deserialize, Serialize };
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 /// Struct for storing player related data.
